Fix cutscene run-in-place: sharedData_800D32A0_0_s02 was u8 (truncated moveSpeed)
Root cause of "Harry runs in place" in waypoint-walk cutscenes (Cheryl
footsteps intro, and the walk phase of every cutscene that uses
sharedFunc_800D23EC). The Jun07 merge reverted a prior PC-port fix: the
moveSpeed staging global sharedData_800D32A0_0_s02 was re-declared as a
scalar u8 (both the extern in bodyprog.h and the storage in data_stubs.c).

The cutscene anim function (sharedFunc_800CDAA8_0_s02) stages the player's
move speed through this global: PlayerState_Unk54 ramps moveSpeed to
Q12(4.0) = 0x4000, writes it to sharedData_800D32A0_0_s02, then reads it
back into playerChara->moveSpeed at the apply block. As a u8, 0x4000
truncates to 0x00 -> moveSpeed reads 0 -> position never advances while the
run animation plays. (The sibling turn-delta global sharedData_800E39D8_0_s00
is correctly s32, which is why the turn-in-place phase worked but the walk
did not.) Pre-merge the storage stub was u8[256] with the header accessing
the first 4 bytes as s32; restore the correct s32 type for both decl and
storage.

+++ include/bodyprog/bodyprog.h
@@ -1025,7 +1025,7 @@ extern u8      D_800AF220;
 extern s16     D_800C42D0;
 extern s16     D_800C42D2;
 extern u8      D_800C4606;
-extern u8      sharedData_800D32A0_0_s02;
+extern s32     sharedData_800D32A0_0_s02;
 extern s32     g_Player_LastWeaponSelected;
 extern s8      g_Player_RockDrill_DirectionAttack;
 extern s_AnimInfo HARRY_BASE_ANIM_INFOS[256];
